+++ src/GRA.Controllers/JoinController.cs
@@ -111,7 +111,6 @@ namespace GRA.Controllers
             var systemList = await _siteService.GetSystemList();
             var programList = await _siteService.GetProgramList();
             var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
-            var districtList = await _schoolService.GetDistrictsAsync(true);
 
             var viewModel = new SinglePageViewModel
             {
@@ -119,11 +118,7 @@ namespace GRA.Controllers
                 ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject),
                 SystemList = new SelectList(systemList.ToList(), "Id", "Name"),
                 ProgramList = new SelectList(programList.ToList(), "Id", "Name"),
-                CategorySelectionAction = nameof(SchoolCategory),
-                PublicSelected = true,
-                ShowPrivateOption = await _schoolService.AnyPrivateSchoolsAsync(),
-                ShowCharterOption = await _schoolService.AnyCharterSchoolsAsync(),
-                SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name")
+                SchoolList = new SelectList(await _schoolService.GetSchoolsAsync(), "Id", "Name")
             };
 
             var askIfFirstTime = await GetSiteSettingBoolAsync(SiteSettingKey.Users.AskIfFirstTime);
@@ -153,7 +148,7 @@ namespace GRA.Controllers
 
             if (systemList.Count() == 1)
             {
-                var systemId = systemList.SingleOrDefault().Id;
+                var systemId = systemList.Single().Id;
                 var branchList = await _siteService.GetBranches(systemId);
                 if (branchList.Count() > 1)
                 {
@@ -161,7 +156,7 @@ namespace GRA.Controllers
                 }
                 else
                 {
-                    viewModel.BranchId = branchList.SingleOrDefault().Id;
+                    viewModel.BranchId = branchList.SingleOrDefault()?.Id;
                 }
                 viewModel.BranchList = new SelectList(branchList.ToList(), "Id", "Name");
                 viewModel.SystemId = systemId;
@@ -169,21 +164,11 @@ namespace GRA.Controllers
 
             if (programList.Count() == 1)
             {
-                var programId = programList.SingleOrDefault().Id;
+                var programId = programList.Single().Id;
                 var program = await _siteService.GetProgramByIdAsync(programId);
-                viewModel.ProgramId = programList.SingleOrDefault().Id;
+                viewModel.ProgramId = programId;
                 viewModel.ShowAge = program.AskAge;
                 viewModel.ShowSchool = program.AskSchool;
-            }
-
-            if (districtList.Count == 1)
-            {
-                viewModel.SchoolDistrictId = districtList.SingleOrDefault().Id;
-                var typeList = await _schoolService.GetTypesAsync(viewModel.SchoolDistrictId);
-                viewModel.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name");
-                var schoolList = await _schoolService.GetSchoolsAsync(viewModel.SchoolDistrictId,
-                    viewModel.SchoolTypeId);
-                viewModel.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
             }
 
             if (useAuthCode)
@@ -192,111 +177,6 @@ namespace GRA.Controllers
             }
 
             return View(viewModel);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> SchoolCategory(SinglePageViewModel model)
-        {
-            ModelState.Clear();
-
-            model.PublicSelected = false;
-            model.PrivateSelected = false;
-            model.CharterSelected = false;
-            model.IsHomeschooled = false;
-            if (model.SetPrivate)
-            {
-                model.PrivateSelected = true;
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetPrivateSchoolListAsync(), "Id", "Name");
-                model.SchoolDistrictId = null;
-            }
-            else if (model.SetCharter)
-            {
-                model.CharterSelected = true;
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetCharterSchoolListAsync(), "Id", "Name");
-                model.SchoolDistrictId = null;
-            }
-            else if (model.SetHomeschool)
-            {
-                model.IsHomeschooled = true;
-                model.SchoolDistrictId = null;
-            }
-            else
-            {
-                model.PublicSelected = true;
-                var districtList = await _schoolService.GetDistrictsAsync(true);
-                model.SchoolDistrictList = new SelectList(districtList, "Id", "Name");
-                if (districtList.Count == 1)
-                {
-                    model.SchoolDistrictId = districtList.SingleOrDefault().Id;
-                    var typeList = await _schoolService.GetTypesAsync(model.SchoolDistrictId);
-                    model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name");
-                    var schoolList = await _schoolService.GetSchoolsAsync(model.SchoolDistrictId);
-                    model.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
-                }
-                else
-                {
-                    model.SchoolDistrictId = null;
-                }
-            }
-
-            model.CategorySelectionAction = nameof(SchoolCategory);
-            model.SchoolId = null;
-            model.SchoolNotListed = false;
-
-            var site = await GetCurrentSiteAsync();
-            PageTitle = $"{site.Name} - Join Now!";
-
-            if (model.SystemId.HasValue)
-            {
-                var branchList = await _siteService.GetBranches(model.SystemId.Value);
-                if (model.BranchId < 1)
-                {
-                    branchList = branchList.Prepend(new Branch() { Id = -1 });
-                }
-                model.BranchList = new SelectList(branchList.ToList(), "Id", "Name");
-            }
-            var systemList = await _siteService.GetSystemList();
-            var programList = await _siteService.GetProgramList();
-            var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
-            model.SystemList = new SelectList(systemList.ToList(), "Id", "Name");
-            model.ProgramList = new SelectList(programList.ToList(), "Id", "Name");
-            model.ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject);
-            model.RequirePostalCode = site.RequirePostalCode;
-
-            if (model.ProgramId.HasValue)
-            {
-                var program = await _siteService.GetProgramByIdAsync(model.ProgramId.Value);
-                model.ShowAge = program.AskAge;
-                model.ShowSchool = program.AskSchool;
-            }
-
-            var askIfFirstTime = await GetSiteSettingBoolAsync(SiteSettingKey.Users.AskIfFirstTime);
-            if (askIfFirstTime)
-            {
-                model.AskFirstTime = EmptyNoYes();
-            }
-
-            var (askEmailSubscription, askEmailSubscriptionText) = await GetSiteSettingStringAsync(
-                SiteSettingKey.Users.AskEmailSubPermission);
-            if (askEmailSubscription)
-            {
-                model.AskEmailSubscription = EmptyNoYes();
-                model.AskEmailSubscriptionText = askEmailSubscriptionText;
-            }
-
-            var (askActivityGoal, defaultDailyGoal) = await GetSiteSettingIntAsync(
-                SiteSettingKey.Users.DefaultDailyPersonalGoal);
-            if (askActivityGoal)
-            {
-                var pointTranslation = programList.First().PointTranslation;
-                model.TranslationDescriptionPastTense =
-                    pointTranslation.TranslationDescriptionPastTense.Replace("{0}", "").Trim();
-                model.ActivityDescriptionPlural = pointTranslation.ActivityDescriptionPlural;
-            }
-
-            return View(nameof(Index), model);
         }
 
         [HttpPost]
@@ -350,12 +230,6 @@ namespace GRA.Controllers
                 if (program.AgeRequired && !model.Age.HasValue)
                 {
                     ModelState.AddModelError("Age", "The Age field is required.");
-                }
-                if (program.SchoolRequired && !model.SchoolId.HasValue
-                    && !model.SchoolDistrictId.HasValue && !model.SchoolNotListed
-                    && !model.IsHomeschooled)
-                {
-                    ModelState.AddModelError("SchoolDistrictId", "The School District field is required.");
                 }
                 if (program.SchoolRequired && !model.SchoolId.HasValue && !model.SchoolNotListed
                     && !model.IsHomeschooled)
@@ -489,70 +363,11 @@ namespace GRA.Controllers
             var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
             model.SystemList = new SelectList(systemList.ToList(), "Id", "Name");
             model.ProgramList = new SelectList(programList.ToList(), "Id", "Name");
+            model.SchoolList = new SelectList(await _schoolService.GetSchoolsAsync(), "Id", "Name");
             model.ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject);
             model.RequirePostalCode = site.RequirePostalCode;
             model.ShowAge = askAge;
             model.ShowSchool = askSchool;
-
-            model.CategorySelectionAction = nameof(SchoolCategory);
-            var districtList = await _schoolService.GetDistrictsAsync(true);
-            if (model.PrivateSelected)
-            {
-                model.PublicSelected = false;
-                model.CharterSelected = false;
-                model.IsHomeschooled = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetPrivateSchoolListAsync(), "Id", "Name");
-            }
-            else if (model.CharterSelected)
-            {
-                model.PublicSelected = false;
-                model.PrivateSelected = false;
-                model.IsHomeschooled = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetCharterSchoolListAsync(), "Id", "Name");
-            }
-            else if (model.IsHomeschooled)
-            {
-                model.PublicSelected = false;
-                model.PrivateSelected = false;
-                model.CharterSelected = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-            }
-            else
-            {
-                model.PublicSelected = true;
-                model.PrivateSelected = false;
-                model.CharterSelected = false;
-                model.IsHomeschooled = false;
-
-                if (model.SchoolId.HasValue)
-                {
-                    var schoolDetails = await _schoolService.GetSchoolDetailsAsync(model.SchoolId.Value);
-                    var typeList = await _schoolService.GetTypesAsync(schoolDetails.SchoolDistrictId);
-                    model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name",
-                        schoolDetails.SchoolDistrictId);
-                    model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
-                        schoolDetails.SchoolTypeId);
-                    model.SchoolList = new SelectList(schoolDetails.Schools.ToList(), "Id", "Name");
-                    ModelState.Remove(nameof(model.SchoolTypeId));
-                }
-                else
-                {
-                    model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                    if (model.SchoolDistrictId.HasValue)
-                    {
-                        var typeList = await _schoolService.GetTypesAsync(model.SchoolDistrictId);
-                        model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
-                            model.SchoolTypeId);
-                        var schoolList = await _schoolService.GetSchoolsAsync(model.SchoolDistrictId,
-                            model.SchoolTypeId);
-                        model.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
-                    }
-                }
-            }
 
             if (askIfFirstTime)
             {
@@ -715,17 +530,12 @@ namespace GRA.Controllers
 
             var programList = await _siteService.GetProgramList();
             var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
-            var districtList = await _schoolService.GetDistrictsAsync(true);
 
             var viewModel = new Step2ViewModel
             {
-                CategorySelectionAction = nameof(Step2SchoolCategory),
-                PublicSelected = true,
-                ShowPrivateOption = await _schoolService.AnyPrivateSchoolsAsync(),
-                ShowCharterOption = await _schoolService.AnyCharterSchoolsAsync(),
                 ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject),
                 ProgramList = new SelectList(programList.ToList(), "Id", "Name"),
-                SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name")
+                SchoolList = new SelectList(await _schoolService.GetSchoolsAsync(), "Id", "Name")
             };
 
             if (programList.Count() == 1)
@@ -737,85 +547,7 @@ namespace GRA.Controllers
                 viewModel.ShowSchool = program.AskSchool;
             }
 
-            if (districtList.Count == 1)
-            {
-                viewModel.SchoolDistrictId = districtList.SingleOrDefault().Id;
-                var typeList = await _schoolService.GetTypesAsync(viewModel.SchoolDistrictId);
-                viewModel.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name");
-                var schoolList = await _schoolService.GetSchoolsAsync(viewModel.SchoolDistrictId,
-                    viewModel.SchoolTypeId);
-                viewModel.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
-            }
-
             return View(viewModel);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Step2SchoolCategory(Step2ViewModel model)
-        {
-            ModelState.Clear();
-
-            model.PublicSelected = false;
-            model.PrivateSelected = false;
-            model.CharterSelected = false;
-            model.IsHomeschooled = false;
-            if (model.SetPrivate)
-            {
-                model.PrivateSelected = true;
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetPrivateSchoolListAsync(), "Id", "Name");
-                model.SchoolDistrictId = null;
-            }
-            else if (model.SetCharter)
-            {
-                model.CharterSelected = true;
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetCharterSchoolListAsync(), "Id", "Name");
-                model.SchoolDistrictId = null;
-            }
-            else if (model.SetHomeschool)
-            {
-                model.IsHomeschooled = true;
-                model.SchoolDistrictId = null;
-            }
-            else
-            {
-                model.PublicSelected = true;
-                var districtList = await _schoolService.GetDistrictsAsync(true);
-                model.SchoolDistrictList = new SelectList(districtList, "Id", "Name");
-                if (districtList.Count == 1)
-                {
-                    model.SchoolDistrictId = districtList.SingleOrDefault().Id;
-                    var typeList = await _schoolService.GetTypesAsync(model.SchoolDistrictId);
-                    model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name");
-                    var schoolList = await _schoolService.GetSchoolsAsync(model.SchoolDistrictId);
-                    model.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
-                }
-                else
-                {
-                    model.SchoolDistrictId = null;
-                }
-            }
-
-            model.CategorySelectionAction = nameof(Step2SchoolCategory);
-            model.SchoolId = null;
-            model.SchoolNotListed = false;
-
-            var site = await GetCurrentSiteAsync();
-            PageTitle = $"{site.Name} - Join Now!";
-
-            var programList = await _siteService.GetProgramList();
-            var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
-            model.ProgramList = new SelectList(programList.ToList(), "Id", "Name");
-            model.ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject);
-            if (model.ProgramId.HasValue)
-            {
-                var program = await _siteService.GetProgramByIdAsync(model.ProgramId.Value);
-                model.ShowAge = program.AskAge;
-                model.ShowSchool = program.AskSchool;
-            }
-
-            return View(nameof(Step2), model);
         }
 
         [HttpPost]
@@ -841,12 +573,6 @@ namespace GRA.Controllers
                 if (program.AgeRequired && !model.Age.HasValue)
                 {
                     ModelState.AddModelError("Age", "The Age field is required.");
-                }
-                if (program.SchoolRequired && !model.SchoolId.HasValue
-                    && !model.SchoolDistrictId.HasValue && !model.SchoolNotListed
-                    && !model.IsHomeschooled)
-                {
-                    ModelState.AddModelError("SchoolDistrictId", "The School District field is required.");
                 }
                 if (program.SchoolRequired && !model.SchoolId.HasValue && !model.SchoolNotListed
                     && !model.IsHomeschooled)
@@ -886,69 +612,10 @@ namespace GRA.Controllers
             var programList = await _siteService.GetProgramList();
             var programViewObject = _mapper.Map<List<ProgramSettingsViewModel>>(programList);
             model.ProgramList = new SelectList(programList.ToList(), "Id", "Name");
+            model.SchoolList = new SelectList(await _schoolService.GetSchoolsAsync(), "Id", "Name");
             model.ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject);
             model.ShowAge = askAge;
             model.ShowSchool = askSchool;
-
-            model.CategorySelectionAction = nameof(Step2SchoolCategory);
-            var districtList = await _schoolService.GetDistrictsAsync(true);
-            if (model.PrivateSelected)
-            {
-                model.PublicSelected = false;
-                model.CharterSelected = false;
-                model.IsHomeschooled = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetPrivateSchoolListAsync(), "Id", "Name");
-            }
-            else if (model.CharterSelected)
-            {
-                model.PublicSelected = false;
-                model.PrivateSelected = false;
-                model.IsHomeschooled = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                model.SchoolList = new SelectList(
-                    await _schoolService.GetCharterSchoolListAsync(), "Id", "Name");
-            }
-            else if (model.IsHomeschooled)
-            {
-                model.PublicSelected = false;
-                model.PrivateSelected = false;
-                model.CharterSelected = false;
-                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-            }
-            else
-            {
-                model.PublicSelected = true;
-                model.PrivateSelected = false;
-                model.CharterSelected = false;
-                model.IsHomeschooled = false;
-
-                if (model.SchoolId.HasValue)
-                {
-                    var schoolDetails = await _schoolService.GetSchoolDetailsAsync(model.SchoolId.Value);
-                    var typeList = await _schoolService.GetTypesAsync(schoolDetails.SchoolDistrictId);
-                    model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name",
-                        schoolDetails.SchoolDistrictId);
-                    model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
-                        schoolDetails.SchoolTypeId);
-                    model.SchoolList = new SelectList(schoolDetails.Schools.ToList(), "Id", "Name");
-                    ModelState.Remove(nameof(model.SchoolTypeId));
-                }
-                else
-                {
-                    model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
-                    if (model.SchoolDistrictId.HasValue)
-                    {
-                        var typeList = await _schoolService.GetTypesAsync(model.SchoolDistrictId);
-                        model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
-                            model.SchoolTypeId);
-                        var schoolList = await _schoolService.GetSchoolsAsync(model.SchoolDistrictId,
-                            model.SchoolTypeId);
-                        model.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
-                    }
-                }
-            }
 
             return View(model);
         }