+++ src/GRA.Data/Model/SchoolDistrict.cs
@@ -12,9 +12,6 @@ namespace GRA.Data.Model
         [MaxLength(255)]
         public string Name { get; set; }
 
-        public bool IsCharter { get; set; }
-        public bool IsPrivate { get; set; }
-
         public virtual ICollection<School> Schools { get; set; }
     }
 }