+++ src/GRA.Domain.Model/SchoolImportExport.cs
@@ -2,7 +2,6 @@
 {
     public class SchoolImportExport
     {
-        public string Type { get; set; }
         public string District { get; set; }
         public string Name { get; set; }
     }