+++ src/GRA.Domain.Repository/ISchoolRepository.cs
@@ -7,9 +7,7 @@ namespace GRA.Domain.Repository
 {
     public interface ISchoolRepository : IRepository<School>
     {
-        Task<ICollection<School>> GetAllAsync(int siteId,
-            int? districtId = default(int?),
-            int? typeId = default(int?));
+        Task<ICollection<School>> GetAllAsync(int siteId, int? districtId = default(int?));
 
         Task<bool> IsInUseAsync(int siteId, int schoolId);
 
@@ -17,10 +15,5 @@ namespace GRA.Domain.Repository
         Task<int> CountAsync(BaseFilter filter);
 
         Task<bool> ValidateAsync(int schoolId, int siteId);
-
-        Task<bool> AnyPrivateSchoolsAsync(int siteId);
-        Task<List<School>> GetPrivateSchoolListAsync(int siteId);
-        Task<bool> AnyCharterSchoolsAsync(int siteId);
-        Task<List<School>> GetCharterSchoolListAsync(int siteId);
     }
 }